Update analytics.py
I converted the primitive units to egohoods by replacing the specific attributes of each primitive units by its neighbor's sum, using the spatial lag operator and the W class from libpysal.weight module. However, in this approach the new attribute is the sum of all neighbor's attirbutes. Do we need to consider the weight? In this modification, all the neighbors have weight 1, we could build the weights on different criteria: contiguity weight, distance-based weight etc. Is it possible to consider the weight when we add the "convert_egohood" option?